+++ dotnet/Apps/Database/Domain.Tests/Order/RequestTests.cs
@@ -104,6 +104,23 @@ namespace Allors.Database.Domain.Tests
 
             Assert.Contains("changedaddress", request.SearchString);
         }
+
+        [Fact]
+        public void ChangedFullfillContactMechanismDisplayNameDeriveRequestForQuoteSearchString()
+        {
+            var webAddress = new WebAddressBuilder(this.Transaction).WithElectronicAddressString("originaladdress").Build();
+            var request = new RequestForQuoteBuilder(this.Transaction)
+                .WithFullfillContactMechanism(webAddress)
+                .Build();
+            this.Transaction.Derive();
+
+            Assert.Contains("originaladdress", request.SearchString);
+
+            webAddress.ElectronicAddressString = "changedaddress";
+            this.Transaction.Derive();
+
+            Assert.Contains("changedaddress", request.SearchString);
+        }
     }
 
     public class RequestAnonymousRuleTests : DomainTest, IClassFixture<Fixture>